Update Bacterial Kidney Disease (BKD).md
check ref 7, as it is paywalled and i can not get around or through it.
ref 8 is norwegian language, but the info is useful or indispensable, depending on the use in the text.
the meta tag #Salmonoids is incorrect, and should be changed to #Salmonids (salmonoid implies something is like a salmon in form, Salmonid is nomenclature for a distinct group).
the section Citations is formatted as an incomplete References section, and a complete reference is made in the "Latest Research Findings" section only for select citations -- i do not intend to decide formatting, but it is incomplete to cite something without a complete formatted reference in a designated section. this would lead to definite blowback from academics, trust me. we should, if possible, choose a distinct referencing style and stick to it (i suggest APA because i am experienced with it).
an additional note on citations: the in-text numbering of these is very useful and wiki-style, though they are all placed at the end of subsections of text, rather than truly in-line with where they are relevant. this is not necessarily wrong, but again, will be picked apart by academics.

overall, a handful of dated references, where some citations are incorrectly "using the text" or have since been made redundant/obsolete/in need of inclusion of supporting evidence, but otherwise a decent section -- made better now =')